Intermediate option added to force subsetting even if no_correlation=1
This introduces a boolean variable `subset_overlap` which is currently hard-coded to be 1. This is a quick way to force the subsetting to use overlapping models for zos and zostoga and thus to have the same ensemble when no_correlation=0. Ideally, we want to add this as a uderdefined option like
```
parser.add_argument('--subset_overlap', help="Subset the models to overlap even when no_correlation=1", type=int, choices=[0,1], default=0)
```