Quick fix to prevent broadcasting issue
Added a conditional check in the Smooth function to return the input array unchanged if its length is smaller than the smoothing window. This fix should address the ValueError encountered due to operand broadcasting with shapes that do not match when the input array length is smaller than the smoothing window. The error occurs in the rare case for example when model = 'bcc-csm1-1' and location='KUNGSHOLMSFORT'