Extrapolate ZOSTOGA and ZOS beyond 2100 to accommodate 19-year smoothing
Implements an extrapolation for the ZOSTOGA and ZOS time series beyond the year 2100. Most models only provide data till 2100. This means that the 19-year average at 2100 is represented by 2090-2100 and is not a true 19-year average. This is why the smoothed time series often exhibit strange ""hook-line" features at the end, which is not real and only because of how the smoothing was done. In order to facilitate a more representative average at 2100, the extrapolation is applied for model time series that do not extend beyond 2110.

The extrapolation is performed using a linear regression based on the last 30 years of valid data, and the additional 10 years are appended to the original ZOSTOGA and ZOS numpy arrays.